Add arg checking to the clock sync kill

+++ hpt/clock_sync_kill.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ $# -ne 1 ]]; then
+	echo "Usage: $0 <hpt card name>"
+	exit 1
+fi
+
 pkill exanic-clock-sy
 pushd /root/jcw78/scripts/hpt_setup/exanic-software/util/
 # Make sure that the PPS master pulse is off
